// src/views/multi-cloud/cloud-host/components/interface.ts
import { IHooksOptions } from '@/hooks/interface'
export interface IHooksBasic {
  billingModeList?: any[]
  instanceList?: any[]
  subsceneList?: any[]
  sceneList?: any[]
  cpuArchitectureList?: any[]
  specsTypeList?: any[]
  instanceSpecsList?: any[]
  mirrorTypeList?: any[]
  systemList?: any[]
  mirrorList?: any[]
  systemDiskList?: any[]
  dataDiskList?: any[]
  dataDiskQuota?: number
  filter?: any
  tableState?: IHooksOptions
}

export interface IHooksNetwork {
  vpcList?: any[]
  subnetList?: any[]
  expandSubnetList?: any[]
  networkCardList?: any[]
  ipGenerateList?: any[]
  safeGroupList?: any[]
  mergeData?: any[]
  lineList?: any[]
  exitEipList?: any[]
  bandwidthTypeList?: any[]
  flowBandWidthList?: any[]
  bandWidthList?: any[]
  shareBandwidth?: any[]
  tableState?: IHooksOptions
  form?: any
}

export interface IHooksHigh {
  loginCredentialsList?: any[]
  keyPairList?: any[]
  cloudBackupList?: any[]
  cloudBackupRepositoryUnit?: any[]
  cloudHostGroups?: any[]
  backupPolicyList?: any[]
  backupRepositoryList?: any[]
  form?: any
}